+++ lib/src/manager/video_manager.dart
@@ -142,6 +142,10 @@ class FlickVideoManager extends ChangeNotifier {
     }
 
     if (autoPlay && ModalRoute.of(_flickManager._context!)!.isCurrent) {
+      //Chrome's autoplay policies are simple:
+      //Muted autoplay is always allowed.
+      if (kIsWeb) _flickManager.flickControlManager!.mute();
+
       // Start playing the video.
       _flickManager.flickControlManager!.play();
     }